compilation champs - repartition hpp cpp
